Moved test-related wip to separate files and put some basic, passing, shield-related assertions in place
